ProjectManager Backend/requirements.txt
asgiref==3.11.1
Django==6.0.2
django-cors-headers==4.9.0
django-rest-framework-mongoengine==3.4.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dnspython==2.8.0
mongoengine==0.29.1
PyJWT==2.11.0
pymongo==4.16.0
python-dotenv==1.2.1
sqlparse==0.5.5
tzdata==2025.3
mongomock
werkzeug
